Add files via upload
Enemy A is the script for the standard enemy.  Has variable stats like health and speed, as well as a lane variable.  The enemy tracks the player, and if the player shoots when the enemy is in the same range, the enemy takes damage.

Spawner0 is one of the individual spawner scripts that determine when and how enemies spawn.  It also tracks how many enemies are in the current lane, so enemies in the back don't get destroyed when enemies in the front get hit.

PM_V2 determines player movement, and moves the avatar around the screen.

Enemies are moving towards the core, and the core controller determines how they operate.  If enemies touch the core too many times, the player loses.